refactor server/utils
