+++ server/handlers/comments/create_comment.go
@@ -3,7 +3,7 @@ package comments
 import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/lareii/copl.uk/server/models"
-	"github.com/lareii/copl.uk/server/utils"
+	"github.com/lareii/copl.uk/server/validate"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
@@ -26,7 +26,7 @@ func CreateComment(c *fiber.Ctx) error {
 		})
 	}
 
-	if err := utils.Validate.Struct(&body); err != nil {
+	if err := validate.Struct(&body); err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
 			"message": "Missing or invalid fields.",
 		})
